AGNO-example-agent/requirements.txt
agno
google-genai
python-dotenv
markitdown[all]
